requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
distro==1.8.0
dj-database-url==2.0.0
Django==4.0.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.5.0
django-jet==1.0.8
django-js-asset==2.1.0
django-rest-passwordreset==1.3.0
django-solo==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
ecdsa==0.18.0
fido2==1.1.1
frozenlist==1.4.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
inflection==0.5.1
intelhex==2.3.0
JPype1==1.4.1
kombu==5.3.2
multidict==6.0.4
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
Pillow==9.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.2.1
pyusb==1.2.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
solo-python==0.1.1
solo1==0.1.1
sqlparse==0.4.2
tabula-py==2.8.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.9
whitenoise==6.2.0
yarl==1.9.4
